Ensure dtype consistency

+++ nn_conditioning_model.py
@@ -145,9 +145,9 @@ class KNNExpertDataset(Dataset):
         final_neighbors = np.argpartition(neighbor_distances, kth=(final_neighbor_num - 1))[:final_neighbor_num]
         neighbor_states = self.flattened_obs_matrix[nearest_neighbors[final_neighbors]]
 
-        return (torch.tensor(neighbor_states, device=device), 
-                torch.tensor(neighbor_distances[final_neighbors], device=device), 
-                torch.tensor(action, device=device))
+        return (torch.tensor(neighbor_states, dtype=torch.float32, device=device), 
+                torch.tensor(neighbor_distances[final_neighbors], dtype=torch.float32, device=device), 
+                torch.tensor(action, dtype=torch.float32, device=device))
 
 def train_model(model, train_loader, num_epochs=100, lr=1e-4):
     criterion = nn.MSELoss()
